# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Adaptive Learning Support App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/firebase-session:
    post:
      operationId: firebaseSession
      tags: [auth]
      summary: Exchange a Firebase Authentication ID token for an app session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirebaseSessionInput"
      responses:
        "200":
          description: Session established (existing user matched by Firebase uid or email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "201":
          description: Session established (new user created from Firebase account)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid or expired Firebase ID token
...
  /dashboard/student/{userId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Student dashboard summary
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Student dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
  /dashboard/parent/{userId}:
    get:
      operationId: getParentDashboard
      tags: [dashboard]
      summary: Parent dashboard — progress of linked children
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Parent dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentDashboard"
  /dashboard/teacher/{userId}:
    get:
      operationId: getTeacherDashboard
      tags: [dashboard]
      summary: Teacher dashboard — all students summary
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Teacher dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDashboard"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [student, parent, teacher, admin]
        avatarUrl:
          type: ["string", "null"]
        gradeLevel:
          type: ["number", "null"]
        age:
          type: ["number", "null"]
        parentId:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - displayName
        - role
        - createdAt
...
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, parent, teacher, admin]
      required:
        - email
        - password
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [student, parent, teacher, admin]
        gradeLevel:
          type: number
        age:
          type: number
      required:
        - email
        - password
        - displayName
        - role
    FirebaseSessionInput:
      type: object
      properties:
        idToken:
          type: string
          description: The Firebase Authentication ID token from the frontend's signed-in user
        # Only used the first time a brand-new Firebase account signs in (no
        # matching user by firebaseUid or email yet). Ignored for existing users.
        displayName:
          type: string
        role:
          type: string
          enum: [student, parent, teacher, admin]
        gradeLevel:
          type: number
        age:
          type: number
      required:
        - idToken
...
    UserProgress:
      type: object
      properties:
        userId:
          type: number
        totalLessonsCompleted:
          type: number
        totalTimeMinutes:
          type: number
        overallAccuracy:
          type: number
        currentStreak:
          type: number
        longestStreak:
          type: number
        level:
          type: number
        xp:
          type: number
        stars:
          type: number
        coins:
          type: number
      required:
        - userId
        - totalLessonsCompleted
        - totalTimeMinutes
        - overallAccuracy
        - currentStreak
        - longestStreak
        - level
        - xp
        - stars
        - coins
    LessonProgress:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        subject:
          type: string
          enum: [math, english, phonics, science, geography, pshe, typing]
        lessonId:
          type: string
        lessonTitle:
          type: string
        score:
          type: number
        accuracy:
          type: number
        timeSpentMinutes:
          type: number
        completedAt:
          type: string
      required:
        - id
        - userId
        - subject
        - lessonId
        - lessonTitle
        - score
        - accuracy
        - timeSpentMinutes
        - completedAt
...
    ProgressDataPoint:
      type: object
      properties:
        date:
          type: string
        lessonsCompleted:
          type: number
        timeMinutes:
          type: number
        accuracy:
          type: number
      required:
        - date
        - lessonsCompleted
        - timeMinutes
        - accuracy
    SubjectProgress:
      type: object
      properties:
        subject:
          type: string
          enum: [math, english, phonics, science, geography, pshe, typing]
        lessonsCompleted:
          type: number
        totalLessons:
          type: number
        accuracy:
          type: number
        timeMinutes:
          type: number
      required:
        - subject
        - lessonsCompleted
        - totalLessons
        - accuracy
        - timeMinutes
...
    Achievement:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          enum: [math, english, phonics, science, geography, pshe, typing, games, streaks, general]
        xpReward:
          type: number
        coinsReward:
          type: number
      required:
        - id
        - title
        - description
        - icon
        - category
        - xpReward
        - coinsReward
    UserAchievement:
      type: object
      properties:
        achievement:
          $ref: "#/components/schemas/Achievement"
        unlockedAt:
          type: string
      required:
        - achievement
        - unlockedAt
    UserRewards:
      type: object
      properties:
        userId:
          type: number
        stars:
          type: number
        coins:
          type: number
        xp:
          type: number
        level:
          type: number
        xpToNextLevel:
          type: number
        unlockedAvatars:
          type: array
          items:
            type: string
        activeAvatar:
          type: string
      required:
        - userId
        - stars
        - coins
        - xp
        - level
        - xpToNextLevel
        - unlockedAvatars
        - activeAvatar
...
    StudentDashboard:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        rewards:
          $ref: "#/components/schemas/UserRewards"
        recentProgress:
          type: array
          items:
            $ref: "#/components/schemas/LessonProgress"
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectProgress"
        recentAchievements:
          type: array
          items:
            $ref: "#/components/schemas/UserAchievement"
        dailyChallengeCompleted:
          type: boolean
        weeklyStreak:
          type: number
      required:
        - user
        - rewards
        - recentProgress
        - subjectBreakdown
        - recentAchievements
        - dailyChallengeCompleted
        - weeklyStreak
    ParentDashboard:
      type: object
      properties:
        children:
          type: array
          items:
            $ref: "#/components/schemas/ChildSummary"
        weeklyActivity:
          type: array
          items:
            $ref: "#/components/schemas/ProgressDataPoint"
      required:
        - children
        - weeklyActivity
    ChildSummary:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        progress:
          $ref: "#/components/schemas/UserProgress"
        recentLessons:
          type: array
          items:
            $ref: "#/components/schemas/LessonProgress"
        subjectBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SubjectProgress"
      required:
        - user
        - progress
        - recentLessons
        - subjectBreakdown
    TeacherDashboard:
      type: object
      properties:
        totalStudents:
          type: number
        avgAccuracy:
          type: number
        activeToday:
          type: number
        topPerformers:
          type: array
          items:
            $ref: "#/components/schemas/StudentSummary"
        studentsNeedingHelp:
          type: array
          items:
            $ref: "#/components/schemas/StudentSummary"
        subjectStats:
          type: array
          items:
            $ref: "#/components/schemas/SubjectStats"
      required:
        - totalStudents
        - avgAccuracy
        - activeToday
        - topPerformers
        - studentsNeedingHelp
        - subjectStats
    StudentSummary:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        accuracy:
          type: number
        lessonsCompleted:
          type: number
        lastActive:
          type: string
      required:
        - user
        - accuracy
        - lessonsCompleted
        - lastActive
    SubjectStats:
      type: object
      properties:
        subject:
          type: string
        avgAccuracy:
          type: number
        totalLessonsCompleted:
          type: number
      required:
        - subject
        - avgAccuracy
        - totalLessonsCompleted
...