# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Adaptive Learning Support App API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputRole } from './registerInputRole';

export interface RegisterInput {
  email: string;
  password: string;
  displayName: string;
  role: RegisterInputRole;
  gradeLevel?: number;
  age?: number;
}